+++ py/polar2grid/polar2grid/grids/config_helper.py
@@ -49,7 +49,7 @@ import os
 import sys
 
 #grid_name,proj4, proj4_str,width,height,pixel_size_x,pixel_size_y,origin_x,origin_y
-CONFIG_LINE_FORMAT = "%(grid_name)s, %(proj4_str)s, %(width)d, %(height)d, %(pixel_size_x)0.3f, %(pixel_size_y)0.3f, %(origin_x)0.3f, %(origin_y)0.3f"
+CONFIG_LINE_FORMAT = "%(grid_name)s, proj4, %(proj4_str)s, %(width)d, %(height)d, %(pixel_size_x)0.3f, %(pixel_size_y)0.3f, %(origin_x)0.3f, %(origin_y)0.3f"
 
 def determine_projection(center_lon, center_lat):
     """Return the 'best' projection to be used based on the center